blender/requirements.txt
autopep8==1.7.0
fake-bpy-module-3.2==20220615
nodeenv==1.7.0
pycodestyle==2.9.1
pydocstyle==6.1.1
pyright==1.1.268
snowballstemmer==2.2.0
toml==0.10.2
